@@ -16,17 +16,18 @@
 package com.google.samples.motionlayoutcodelab
 
 import android.os.Bundle
-import android.view.View
-import com.google.android.material.snackbar.Snackbar
-import androidx.appcompat.app.AppCompatActivity;
 
-import kotlinx.android.synthetic.main.activity_step1.*
+import androidx.appcompat.app.AppCompatActivity
+
+import com.google.samples.motionlayoutcodelab.databinding.ActivityStep1Binding
 
 class Step1Activity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityStep1Binding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_step1)
+        binding = ActivityStep1Binding.inflate(layoutInflater)
+        setContentView(binding.root)
     }
-
 }
