@@ -16,15 +16,14 @@
 package com.google.samples.motionlayoutcodelab
 
 import android.os.Bundle
-import android.view.View
-import android.widget.ImageView
-import com.google.android.material.snackbar.Snackbar
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.lifecycle.LiveData
 
-import kotlinx.android.synthetic.main.activity_step1.*
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.samples.motionlayoutcodelab.databinding.ActivityStep7Binding
 
 class Step7Activity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityStep7Binding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
