@@ -16,31 +16,33 @@
 package com.google.samples.motionlayoutcodelab
 
 import android.os.Bundle
-import android.util.TypedValue
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.constraintlayout.motion.widget.MotionLayout
-import androidx.constraintlayout.widget.ConstraintLayout
+
+import androidx.appcompat.app.AppCompatActivity
+
 import com.google.android.material.appbar.AppBarLayout
+
+import com.google.samples.motionlayoutcodelab.databinding.ActivityStep8CompletedBinding
 
 class Step8CompletedActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityStep8CompletedBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_step8_completed)
+        binding = ActivityStep8CompletedBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
         coordinateMotion()
     }
 
     private fun coordinateMotion() {
-        val appBarLayout: AppBarLayout = findViewById(R.id.appbar_layout)
-        val motionLayout: MotionLayout = findViewById(R.id.motion_layout)
-
-        val listener = AppBarLayout.OnOffsetChangedListener { unused, verticalOffset ->
-            val seekPosition = -verticalOffset / appBarLayout.totalScrollRange.toFloat()
-            motionLayout.progress = seekPosition
+        val listener = AppBarLayout.OnOffsetChangedListener { _, verticalOffset ->
+            val seekPosition = -verticalOffset / binding.appbarLayout.totalScrollRange.toFloat()
+            binding.motionLayout.progress = seekPosition
         }
 
-        appBarLayout.addOnOffsetChangedListener(listener)
+        binding.appbarLayout.addOnOffsetChangedListener(listener)
     }
+
+
 }
